+++ QiLuTreeSeedlings/UI/GongChengGongSiVVVVV/WoDeDingDanViewController/YLDdingdanXiangQingViewController/YLDNvaChangeSearchView/YLDSearchNavView.m
@@ -68,7 +68,7 @@
           
                 if (toBeString.length > kssss) {
                 // NSLog(@"最多%d个字符!!!",kMaxLength);
-                [ToastView showToast:[NSString stringWithFormat:@"最多%ld个字符",kssss] withOriginY:250 withSuperView:self];
+                [ToastView showToast:[NSString stringWithFormat:@"最多%ld个字符",(long)kssss] withOriginY:250 withSuperView:self];
                 //[XtomFunction openIntervalHUD:[NSString stringWithFormat:@"最多%d个字符",kMaxLength] view:nil];
                 textField.text = [toBeString substringToIndex:kssss];
                 if ([self.delegate respondsToSelector:@selector(textFieldChangeVVWithStr:)]) {
@@ -90,7 +90,7 @@
             if (toBeString.length > kssss) {
             //[XtomFunction openIntervalHUD:[NSString stringWithFormat:@"最多%ld个字符",(long)kMaxLength] view:nil];
             //NSLog(@"最多%d个字符!!!",kMaxLength);
-            [ToastView showToast:[NSString stringWithFormat:@"最多%ld个字符",kssss] withOriginY:250 withSuperView:self];
+            [ToastView showToast:[NSString stringWithFormat:@"最多%ld个字符",(long)kssss] withOriginY:250 withSuperView:self];
             textField.text = [toBeString substringToIndex:kssss];
             if ([self.delegate respondsToSelector:@selector(textFieldChangeVVWithStr:)]) {
                     [self.delegate textFieldChangeVVWithStr:textField.text];
